Move color_spaces to util
